requirements.txt
appdirs==1.4.0
arrow==0.8.0
attrs==16.1.0
baidu-aip==1.6.0.0
beautifulsoup4==4.5.1
binaryornot==0.4.0
boto==2.42.0
bz2file==0.98
cffi==1.8.2
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
cookiejar==0.0.1
coreapi==2.3.0
coreschema==0.0.4
coverage==4.2
cryptography==1.5
cssselect==0.9.2
cycler==0.10.0
Django==1.10.1
django-bootstrap-toolkit==2.15.0
django-common-helpers==0.9.0
django-cors-headers==2.0.2
django-cron==0.5.0
django-crontab==0.7.1
django-filter==1.0.1
django-guardian==1.4.6
django-redis==4.7.0
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.1.2
djangorestframework==3.6.2
enum34==1.1.6
funcsigs==1.0.2
future==0.15.2
gensim==0.13.2
http==0.2
idna==2.1
ipaddress==1.0.16
itypes==1.1.0
Jinja2==2.8
jinja2-time==0.2.0
lxml==3.6.4
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.5.3
mock==2.0.0
mongoengine==0.13.0
mysql-connector-python-cext==2.1.3
MySQL-python==1.2.5
numpy==1.11.2
openapi-codec==1.3.1
packaging==16.8
pager==3.3
pandas==0.19.0
parsel==1.0.3
pbr==1.10.0
Pillow==3.3.1
poyo==0.4.0
protobuf==3.1.0
py==1.4.32
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
Pygments==2.2.0
pymongo==3.4.0
pyOpenSSL==16.1.0
pyparsing==2.1.10
pypinyin==0.21.0
pytest==3.0.6
python-dateutil==2.5.3
pytz==2016.7
qiniu==7.1.0
queuelib==1.4.2
redis==2.10.5
requests==2.11.1
rsa==3.4.2
scikit-learn==0.18
scipy==0.18.1
Scrapy==1.1.2
selenium==2.53.6
Serializer==0.2.1
service-identity==16.0.0
simplejson==3.10.0
six==1.10.0
smart-open==1.3.5
tensorflow==0.12.0rc0
Twisted==16.4.0
uritemplate==3.0.0
w3lib==1.15.0
whichcraft==0.4.0
zope.interface==4.3.2
